WIP: Update base Account types & EVM/AVAX, added signature verification functions
