+++ packages/account/src/provider.ts
@@ -3,6 +3,10 @@
  */
 export abstract class BaseProviderWallet {
   /**
+   * Provider property
+   */
+  public readonly provider: unknown
+  /**
    * Connects the Provider and fetch an account
    */
   abstract connect(): Promise<void>
