+++ packages/avalanche/src/account.ts
@@ -1,8 +1,9 @@
 import { Avalanche, BinTools, Buffer as AvaBuff } from 'avalanche'
-import { KeyPair } from 'avalanche/dist/apis/avm'
+import { KeyPair, KeyChain } from 'avalanche/dist/apis/avm'
+import { KeyPair as EVMKeyPair } from 'avalanche/dist/apis/evm'
 import { decrypt as secp256k1_decrypt, encrypt as secp256k1_encrypt } from 'eciesjs'
-import { providers } from 'ethers'
-import shajs from 'sha.js'
+import { ethers, providers } from 'ethers'
+import { privateToAddress } from 'ethereumjs-util'
 
 import { Blockchain } from '@aleph-sdk/core'
 import { SignableMessage, ECIESAccount, BaseProviderWallet } from '@aleph-sdk/account'
@@ -10,28 +11,34 @@ import {
   ProviderEncryptionLabel,
   ProviderEncryptionLib,
   ChangeRpcParam,
+  RpcId,
+  EVMAccount,
   JsonRPCWallet,
-  RpcChainType,
 } from '@aleph-sdk/evm'
+import { digestMessage, verifyAvalanche } from './verify'
 
 /**
  * AvalancheAccount implements the Account class for the Avalanche protocol.
  * It is used to represent an Avalanche account when publishing a message on the Aleph network.
  */
-export class AvalancheAccount extends ECIESAccount {
-  private signer?: KeyPair
-  private provider?: BaseProviderWallet
+export class AvalancheAccount extends EVMAccount {
+  public declare readonly wallet?: BaseProviderWallet
+  public readonly keyPair?: KeyPair | EVMKeyPair
 
-  constructor(signerOrProvider: KeyPair | BaseProviderWallet, address: string, publicKey?: string) {
+  constructor(
+    signerOrWallet: KeyPair | EVMKeyPair | BaseProviderWallet | ethers.providers.JsonRpcProvider,
+    address: string,
+    publicKey?: string,
+  ) {
     super(address, publicKey)
-
-    if (signerOrProvider instanceof KeyPair) this.signer = signerOrProvider
-    if (signerOrProvider instanceof BaseProviderWallet) this.provider = signerOrProvider
+    if (signerOrWallet instanceof ethers.providers.JsonRpcProvider) this.wallet = new JsonRPCWallet(signerOrWallet)
+    else if (signerOrWallet instanceof KeyPair || signerOrWallet instanceof EVMKeyPair) this.keyPair = signerOrWallet
+    else this.wallet = signerOrWallet
   }
 
   override GetChain(): Blockchain {
-    if (this.signer) return Blockchain.AVAX
-    if (this.provider) return Blockchain.ETH
+    if (this.keyPair) return Blockchain.AVAX
+    if (this.wallet) return Blockchain.ETH
 
     throw new Error('Cannot determine chain')
   }
@@ -46,9 +53,8 @@ export class AvalancheAccount extends ECIESAccount {
    */
   override async askPubKey(): Promise<void> {
     if (this.publicKey) return
-    if (!this.provider) throw Error('PublicKey Error: No providers are setup')
-
-    this.publicKey = await this.provider.getPublicKey()
+    if (!this.wallet) throw Error('PublicKey Error: No providers are setup')
+    this.publicKey = await this.wallet.getPublicKey()
     return
   }
 
@@ -80,7 +86,7 @@ export class AvalancheAccount extends ECIESAccount {
     }
 
     if (!publicKey) throw new Error('Cannot encrypt content')
-    if (!this.provider) {
+    if (!this.wallet) {
       // Wallet encryption method or non-metamask provider
       return secp256k1_encrypt(publicKey, content)
     } else {
@@ -95,24 +101,15 @@ export class AvalancheAccount extends ECIESAccount {
    * @param encryptedContent The encrypted content to decrypt.
    */
   async decrypt(encryptedContent: Buffer | string): Promise<Buffer> {
-    if (this.signer) {
-      const secret = this.signer.getPrivateKey().toString('hex')
+    if (this.keyPair) {
+      const secret = this.keyPair.getPrivateKey().toString('hex')
       return secp256k1_decrypt(secret, Buffer.from(encryptedContent))
     }
-    if (this.provider) {
-      const decrypted = await this.provider.decrypt(encryptedContent)
+    if (this.wallet) {
+      const decrypted = await this.wallet.decrypt(encryptedContent)
       return Buffer.from(decrypted)
     }
     throw new Error('Cannot encrypt content')
-  }
-
-  private async digestMessage(message: Buffer) {
-    const msgSize = Buffer.alloc(4)
-    msgSize.writeUInt32BE(message.length, 0)
-    const msgStr = message.toString('utf-8')
-    const msgBuf = Buffer.from(`\x1AAvalanche Signed Message:\n${msgSize}${msgStr}`, 'utf8')
-
-    return new shajs.sha256().update(msgBuf).digest()
   }
 
   /**
@@ -129,32 +126,42 @@ export class AvalancheAccount extends ECIESAccount {
 
     const buffer = message.GetVerificationBuffer()
 
-    const digest = await this.digestMessage(buffer)
+    const digest = digestMessage(buffer)
 
-    if (this.signer) {
+    if (this.keyPair) {
       const digestHex = digest.toString('hex')
       const digestBuff = AvaBuff.from(digestHex, 'hex')
-      const signatureBuffer = this.signer?.sign(digestBuff)
+      const signatureBuffer = this.keyPair?.sign(digestBuff)
 
       const bintools = BinTools.getInstance()
-      return bintools.cb58Encode(signatureBuffer)
-    } else if (this.provider) {
-      return await this.provider.signMessage(buffer)
+      const signature = bintools.cb58Encode(signatureBuffer)
+      if (await verifyAvalanche(buffer, signature, this.keyPair.getPublicKey().toString('hex'))) return signature
+
+      throw new Error('Cannot proof the integrity of the signature')
+    } else if (this.wallet) {
+      return await this.wallet.signMessage(buffer)
     }
 
     throw new Error('Cannot sign message')
   }
 }
 
-async function getKeyChain() {
-  const ava = new Avalanche()
-  const xChain = ava.XChain()
-
-  return xChain.keyChain()
+export enum ChainType {
+  C_CHAIN = 'C',
+  X_CHAIN = 'X',
 }
 
-export async function getKeyPair(privateKey?: string): Promise<KeyPair> {
-  const keyChain = await getKeyChain()
+/**
+ * Get Key Chains
+ * @param chain Avalanche chain type: c-chain | x-chain
+ * @returns key chains
+ */
+async function getKeyChain(chain = ChainType.X_CHAIN) {
+  return new KeyChain(new Avalanche().getHRP(), chain)
+}
+
+export async function getKeyPair(privateKey?: string, chain = ChainType.X_CHAIN): Promise<KeyPair> {
+  const keyChain = await getKeyChain(chain)
   const keyPair = keyChain.makeKey()
 
   if (privateKey) {
@@ -172,14 +179,18 @@ export async function getKeyPair(privateKey?: string): Promise<KeyPair> {
 }
 
 /**
- * Imports an Avalanche account given a private key.
+ * Imports an Avalanche account given a private key and chain
  *
  * It creates an Avalanche keypair containing information about the account, extracted in the AvalancheAccount constructor.
  *
  * @param privateKey The private key of the account to import.
+ * @param chain The Avalanche subnet to use the account with.
  */
-export async function ImportAccountFromPrivateKey(privateKey: string): Promise<AvalancheAccount> {
-  const keyPair = await getKeyPair(privateKey)
+export async function ImportAccountFromPrivateKey(
+  privateKey: string,
+  chain = ChainType.X_CHAIN,
+): Promise<AvalancheAccount> {
+  const keyPair = await getKeyPair(privateKey, chain)
   return new AvalancheAccount(keyPair, keyPair.getAddressString(), keyPair.getPublicKey().toString('hex'))
 }
 
@@ -191,7 +202,7 @@ export async function ImportAccountFromPrivateKey(privateKey: string): Promise<A
  */
 export async function GetAccountFromProvider(
   provider: providers.ExternalProvider,
-  requestedRpc: ChangeRpcParam = RpcChainType.AVAX,
+  requestedRpc: ChangeRpcParam = RpcId.AVAX,
 ): Promise<AvalancheAccount> {
   const avaxProvider = new providers.Web3Provider(provider)
   const jrw = new JsonRPCWallet(avaxProvider)
@@ -205,15 +216,38 @@ export async function GetAccountFromProvider(
 }
 
 /**
- * Creates a new Avalanche account using a randomly generated privateKey
+ * Retrieves the EVM compatible address for the current account.
+ * This function works specifically with the C-Chain.
  *
+ * If the current signer is not associated with the C-Chain,
+ * the function throws an error.
+ *
+ * @returns A Promise that resolves to the EVM-style address of the account
+ * @throws An error if the current signer is not associated with the C-Chain
  */
-export async function NewAccount(): Promise<{ account: AvalancheAccount; privateKey: string }> {
-  const keypair = await getKeyPair()
+function getEVMAddress(keypair: EVMKeyPair): string {
+  const pkHex = keypair.getPrivateKey().toString('hex')
+  const pkBuffNative = Buffer.from(pkHex, 'hex')
+  const ethAddress = privateToAddress(pkBuffNative).toString('hex')
+  return `0x${ethAddress}`
+}
+
+/**
+ * Creates a new Avalanche account using a randomly generated privateKey
+ */
+export async function NewAccount(
+  chain = ChainType.X_CHAIN,
+): Promise<{ account: AvalancheAccount; privateKey: string }> {
+  const keypair = await getKeyPair(undefined, chain)
   const privateKey = keypair.getPrivateKey().toString('hex')
 
+  let address: string = keypair.getAddressString()
+  if (chain === ChainType.C_CHAIN) {
+    address = getEVMAddress(keypair)
+  }
+
   return {
-    account: new AvalancheAccount(keypair, keypair.getAddressString(), keypair.getPublicKey().toString('hex')),
+    account: new AvalancheAccount(keypair, address, keypair.getPublicKey().toString('hex')),
     privateKey,
   }
 }
