+++ packages/evm/src/index.ts
@@ -1,1 +1,3 @@
 export * from './provider'
+export * from './account'
+export * from './verify'
