+++ packages/evm/src/provider/rpc.ts
@@ -4,12 +4,13 @@ import { BaseProviderWallet } from '@aleph-sdk/account'
 const RPC_WARNING = `DEPRECATION WARNING: 
 Encryption/Decryption features may become obsolete, for more information: https://github.com/aleph-im/aleph-sdk-ts/issues/37`
 
-export enum RpcChainType {
+export enum RpcId {
   ETH,
   ETH_FLASHBOTS,
   POLYGON,
   BSC,
   AVAX,
+  AVAX_TESTNET,
 }
 
 export type RpcType = {
@@ -24,11 +25,19 @@ export type RpcType = {
   blockExplorerUrls: string[]
 }
 
-export type ChangeRpcParam = RpcType | RpcChainType
+export type ChangeRpcParam = RpcType | RpcId
 
-const ChainData: { [key: string]: RpcType } = {
-  [RpcChainType.AVAX]: {
-    chainId: '0xA86A',
+export function decToHex(dec: number): string {
+  return '0x' + dec.toString(16)
+}
+
+export function hexToDec(hex: string): number {
+  return parseInt(hex.slice(2), 16)
+}
+
+export const ChainData: { [key: number]: RpcType } = {
+  [RpcId.AVAX]: {
+    chainId: decToHex(43114),
     rpcUrls: ['https://api.avax.network/ext/bc/C/rpc'],
     chainName: 'Avalanche Mainnet',
     nativeCurrency: {
@@ -36,10 +45,21 @@ const ChainData: { [key: string]: RpcType } = {
       symbol: 'AVAX',
       decimals: 18,
     },
-    blockExplorerUrls: ['https://snowtrace.io'],
+    blockExplorerUrls: ['https://snowtrace.io/'],
   },
-  [RpcChainType.ETH]: {
-    chainId: '0x1',
+  [RpcId.AVAX_TESTNET]: {
+    chainId: decToHex(43113),
+    rpcUrls: ['https://api.avax-test.network/ext/bc/C/rpc'],
+    chainName: 'Avalanche Testnet',
+    nativeCurrency: {
+      name: 'AVAX',
+      symbol: 'AVAX',
+      decimals: 18,
+    },
+    blockExplorerUrls: ['https://testnet.snowtrace.io/'],
+  },
+  [RpcId.ETH]: {
+    chainId: decToHex(1),
     rpcUrls: ['https://mainnet.infura.io/v3/'],
     chainName: 'Ethereum Mainnet',
     nativeCurrency: {
@@ -49,8 +69,8 @@ const ChainData: { [key: string]: RpcType } = {
     },
     blockExplorerUrls: ['https://etherscan.io'],
   },
-  [RpcChainType.ETH_FLASHBOTS]: {
-    chainId: '0x1',
+  [RpcId.ETH_FLASHBOTS]: {
+    chainId: decToHex(1),
     rpcUrls: ['https://rpc.flashbots.net/'],
     chainName: 'Ethereum Mainnet - Flashbots',
     nativeCurrency: {
@@ -60,8 +80,8 @@ const ChainData: { [key: string]: RpcType } = {
     },
     blockExplorerUrls: ['https://etherscan.io'],
   },
-  [RpcChainType.POLYGON]: {
-    chainId: '0x89',
+  [RpcId.POLYGON]: {
+    chainId: decToHex(137),
     rpcUrls: ['https://polygon-rpc.com/'],
     chainName: 'Polygon Mainnet',
     nativeCurrency: {
@@ -71,8 +91,8 @@ const ChainData: { [key: string]: RpcType } = {
     },
     blockExplorerUrls: ['https://polygonscan.com/'],
   },
-  [RpcChainType.BSC]: {
-    chainId: '0x38',
+  [RpcId.BSC]: {
+    chainId: decToHex(56),
     rpcUrls: ['https://bsc-dataseed.binance.org/'],
     chainName: 'Binance Smart Chain Mainnet',
     nativeCurrency: {
@@ -85,15 +105,15 @@ const ChainData: { [key: string]: RpcType } = {
 }
 
 /**
- * Wrapper for JSON RPC Providers (ex: Metamask)
+ * Wrapper for JSON RPC Providers (ex: Metamask).
  */
 export class JsonRPCWallet extends BaseProviderWallet {
-  private provider: ethers.providers.Web3Provider
+  public declare readonly provider: ethers.providers.JsonRpcProvider | ethers.providers.Web3Provider
   private signer?: ethers.providers.JsonRpcSigner
   public address?: string
   private publicKey?: string
 
-  constructor(provider: ethers.providers.Web3Provider) {
+  constructor(provider: ethers.providers.JsonRpcProvider | ethers.providers.Web3Provider) {
     super()
     this.provider = provider
   }
@@ -135,9 +155,9 @@ export class JsonRPCWallet extends BaseProviderWallet {
     return this.signer.signMessage(data)
   }
 
-  public async changeNetwork(chainOrRpc: RpcType | RpcChainType = RpcChainType.ETH): Promise<void> {
+  public async changeNetwork(chainOrRpc: RpcType | RpcId = RpcId.ETH): Promise<void> {
     if (typeof chainOrRpc === 'number') {
-      if (chainOrRpc === RpcChainType.ETH) {
+      if (chainOrRpc === RpcId.ETH) {
         await this.provider.send('wallet_switchEthereumChain', [{ chainId: '0x1' }])
       } else await this.provider.send('wallet_addEthereumChain', [ChainData[chainOrRpc]])
     } else {
@@ -145,8 +165,13 @@ export class JsonRPCWallet extends BaseProviderWallet {
     }
   }
 
+  public async getCurrentChainId(): Promise<number> {
+    const network = await this.provider.getNetwork()
+    return network.chainId
+  }
+
   public isMetamask(): boolean {
-    return !!this.provider?.provider?.isMetaMask
+    return this.provider instanceof ethers.providers.Web3Provider && !!this.provider?.provider.isMetaMask
   }
 
   public async isConnected(): Promise<boolean> {
