--- src/accounts/providers/Ledger/index.ts
@@ -1,3 +1,0 @@
-import * as ethereum from './ethereum'
-
-export { ethereum }
